bot_controller/requirements.txt

aiogram==3.15.0
aiohttp==3.10.11
pydantic==2.9.2
pydantic-settings==2.7.0
python-dotenv==1.0.1
flake8==7.1.0
mypy==1.10.1
isort==5.13.2
black==24.4.2 
